refactor datepicker and calendar components
