--- packages/ui/src/core/calender/calendar.props.d.ts
@@ -1,33 +1,0 @@
-export interface CalendarProps extends React.HTMLAttributes<HTMLDivElement> {
-  value?: string;
-  _month?: number;
-  className?: string;
-  datepickerInput?: any;
-  onChange?: (value: string | number) => void;
-  open?: boolean;
-  isSelectedDay?: boolean;
-  maxWidth?: boolean;
-  isYearChanged?: boolean;
-  isMonthChanged?: boolean;
-  isNextMonth?: boolean;
-  isPrevMonth?: boolean;
-  isMonthSelected?: boolean;
-  showMonthsDatePicker?: boolean;
-  showYearsDatePicker?: boolean;
-  day?: any;
-  year?: number;
-  _year?: number;
-  monthIndex?: number;
-  currentMonth?: number | string;
-  getMonthStr?: (value: number) => number | string;
-  monthDetails?: array;
-  setMonth?: (value: number) => void;
-  setYear?: (value: number) => void;
-  setSelectedDay?: (value: number) => void;
-  setMonthDetails?: (value: any) => void;
-  setOpenState?: (value: boolean) => void;
-  getMonthDetails?: (value: number, value: number) => object[];
-  getTimeStamp?: () => number;
-  setShowMonthsDatePicker?: (value: boolean) => void;
-  setShowYearsDatePicker?: (value: boolean) => void;
-}
